Restructure files to reduce complexity
